PSR-266 Moved WhoReceivedLoanController into a common package
