+++ conf/app.routes
@@ -280,14 +280,14 @@ GET         /:srn/submit-file-upload-error-missing-information                  
 GET         /:srn/change-file-upload-error-missing-information                          controllers.nonsipp.memberdetails.upload.FileUploadErrorMissingInformationController.onPageLoad(srn: Srn, mode: Mode = CheckMode)
 GET         /:srn/change-submit-file-upload-error-missing-information                   controllers.nonsipp.memberdetails.upload.FileUploadErrorMissingInformationController.onSubmit(srn: Srn, mode: Mode = CheckMode)
 
-GET         /:srn/who-received-loan                                                     controllers.nonsipp.loansmadeoroutstanding.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999 = 1, mode: Mode = NormalMode)
-GET         /:srn/who-received-loan/:index                                              controllers.nonsipp.loansmadeoroutstanding.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999, mode: Mode = NormalMode)
-POST        /:srn/who-received-loan                                                     controllers.nonsipp.loansmadeoroutstanding.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999 = 1, mode: Mode = NormalMode)
-POST        /:srn/who-received-loan/:index                                              controllers.nonsipp.loansmadeoroutstanding.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999, mode: Mode = NormalMode)
-GET         /:srn/change-who-received-loan                                              controllers.nonsipp.loansmadeoroutstanding.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999 = 1, mode: Mode = CheckMode)
-GET         /:srn/change-who-received-loan/:index                                       controllers.nonsipp.loansmadeoroutstanding.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999, mode: Mode = CheckMode)
-POST        /:srn/change-who-received-loan                                              controllers.nonsipp.loansmadeoroutstanding.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999 = 1, mode: Mode = CheckMode)
-POST        /:srn/change-who-received-loan/:index                                       controllers.nonsipp.loansmadeoroutstanding.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999, mode: Mode = CheckMode)
+GET         /:srn/who-received-loan                                                     controllers.nonsipp.common.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999 = 1, mode: Mode = NormalMode)
+GET         /:srn/who-received-loan/:index                                              controllers.nonsipp.common.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999, mode: Mode = NormalMode)
+POST        /:srn/who-received-loan                                                     controllers.nonsipp.common.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999 = 1, mode: Mode = NormalMode)
+POST        /:srn/who-received-loan/:index                                              controllers.nonsipp.common.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999, mode: Mode = NormalMode)
+GET         /:srn/change-who-received-loan                                              controllers.nonsipp.common.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999 = 1, mode: Mode = CheckMode)
+GET         /:srn/change-who-received-loan/:index                                       controllers.nonsipp.common.WhoReceivedLoanController.onPageLoad(srn: Srn, index: Max9999999, mode: Mode = CheckMode)
+POST        /:srn/change-who-received-loan                                              controllers.nonsipp.common.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999 = 1, mode: Mode = CheckMode)
+POST        /:srn/change-who-received-loan/:index                                       controllers.nonsipp.common.WhoReceivedLoanController.onSubmit(srn: Srn, index: Max9999999, mode: Mode = CheckMode)
 
 GET         /:srn/file-upload-error-some                                                controllers.nonsipp.memberdetails.upload.FileUploadErrorSummaryController.onPageLoad(srn: Srn, mode: Mode = NormalMode)
 GET         /:srn/submit-file-upload-error-some                                         controllers.nonsipp.memberdetails.upload.FileUploadErrorSummaryController.onSubmit(srn: Srn, mode: Mode = NormalMode)
